move config detail page to admin
