+++ src/openzaak/config/admin_views.py
@@ -2,7 +2,7 @@ from django.conf import settings
 from django.urls import reverse
 from django.views.generic import TemplateView
 
-from .models import InternalService
+from .models import InternalService, NLXConfig
 from .utils import AdminRequiredMixin
 
 
@@ -17,7 +17,6 @@ class NLXInwayView(AdminRequiredMixin, TemplateView):
             "documentation_url": "https://open-zaak.readthedocs.io/en/latest/",
         }
         version = settings.REST_FRAMEWORK["DEFAULT_VERSION"]
-
         services = []
         for internal_service in InternalService.objects.filter(nlx=True).order_by(
             "api_type"
@@ -41,5 +40,21 @@ class NLXInwayView(AdminRequiredMixin, TemplateView):
             services.append(service)
 
         context.update({"services": services})
+
+        return context
+
+
+class ConfigDetailView(AdminRequiredMixin, TemplateView):
+    template_name = "admin/config_detail.html"
+    login_url = "admin:login"
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+
+        nlx = NLXConfig.get_solo()
+        context["nlx"] = nlx
+
+        internal_services = InternalService.objects.order_by("api_type").all()
+        context["internal_services"] = internal_services
 
         return context
