--- src/openzaak/config/templates/config/config_detail.html
@@ -1,80 +1,0 @@
-{% extends "master.html" %}
-
-{% block content %}
-<div class="px-3 py-3 pt-md-5 pb-md-4 mx-auto text-center">
-    <h1 class="display-4">Service configuratie</h1>
-    <p class="lead">Open Zaak API dashboard</p>
-</div>
-
-<div class="container text-left">
-
-    <div class="text-right">
-        <a href="{% url 'config-nlx' %}" class="btn btn-sm btn-secondary">Wijzigen</a>
-    </div>
-    <!-- NLX config -->
-    <section>
-        <div class="text-center">
-            <h3>NLX</h3>
-        </div>
-
-        <table class="table table-striped">
-            <thead>
-                <tr>
-                    <th scope="col">Parameter</th>
-                    <th scope="col">Waarde</th>
-                </tr>
-            </thead>
-            <tbody>
-                <tr>
-                    <td><strong>NLX directory</strong></td>
-                    <td>{{ nlx.directory_url| urlize }}</td>
-                </tr>
-                <tr>
-                    <td><strong>Outway address</strong></td>
-                    <td>{{ nlx.outway| urlize }}</td>
-                </tr>
-            </tbody>
-        </table>
-    </section>
-
-    <!-- Internal API config -->
-    <section>
-        <div class="text-center">
-            <h3>Open Zaak services</h3>
-        </div>
-
-        <table class="table table-striped">
-            <thead>
-                <tr>
-                    <th scope="col">Service</th>
-                    <th scope="col" class="text-center"
-                        data-toggle="tooltip" title="Service geactiveerd en aangeboden door Open Zaak">
-                        Geactiveerd
-                    </th>
-                    <th scope="col" class="text-center"
-                        data-toggle="tooltip" title="Service wordt aangeboden over NLX">
-                        NLX
-                    </th>
-                </tr>
-            </thead>
-            <tbody>
-                {% for service in internal_services %}
-                <tr>
-                    <td>{{ service.component|title }} API</td>
-                    <td class="text-center text-{{ service.enabled|yesno:'success,danger,muted' }}">
-                        <i class="fa fa-{{ service.enabled|yesno:'check,times,minus' }}"></i>
-                    </td>
-                    <td class="text-center text-{{ service.nlx|yesno:'success,danger,muted' }}">
-                        <i class="fa fa-{{ service.nlx|yesno:'check,times,minus' }}"></i>
-                    </td>
-                </tr>
-                {% endfor %}
-            </tbody>
-        </table>
-    </section>
-
-</div>
-
-
-
-{% endblock %}
