+++ src/openzaak/config/urls.py
@@ -1,9 +1,8 @@
 from django.urls import path
 
-from .views import ConfigDetailView, InternalConfigView, NLXConfigView
+from .views import InternalConfigView, NLXConfigView
 
 urlpatterns = [
-    path("", ConfigDetailView.as_view(), name="config-detail"),
     path("nlx/", NLXConfigView.as_view(), name="config-nlx"),
     path("internal/", InternalConfigView.as_view(), name="config-internal"),
 ]
