+++ src/openzaak/config/views.py
@@ -8,21 +8,6 @@ from .models import InternalService, NLXConfig
 from .utils import AdminRequiredMixin
 
 
-class ConfigDetailView(AdminRequiredMixin, TemplateView):
-    template_name = "config/config_detail.html"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-
-        nlx = NLXConfig.get_solo()
-        context["nlx"] = nlx
-
-        internal_services = InternalService.objects.order_by("api_type").all()
-        context["internal_services"] = internal_services
-
-        return context
-
-
 class NLXConfigView(AdminRequiredMixin, UpdateView):
     model = NLXConfig
     form_class = NLXConfigForm
