+++ src/openzaak/sass/admin/_app_overrides.scss
@@ -43,3 +43,6 @@
   word-wrap: break-word;
 }
 
+.table {
+  width: 100%;
+}
